Deduplicate `enum LrEdgeFlags` to `src/looprestoration.rs`.
