+++ src/lf_apply_tmpl_8.rs
@@ -1147,11 +1147,11 @@ pub type looprestorationfilter_fn = Option::<
         LrEdgeFlags,
     ) -> (),
 >;
-pub type LrEdgeFlags = libc::c_uint;
-pub const LR_HAVE_BOTTOM: LrEdgeFlags = 8;
-pub const LR_HAVE_TOP: LrEdgeFlags = 4;
-pub const LR_HAVE_RIGHT: LrEdgeFlags = 2;
-pub const LR_HAVE_LEFT: LrEdgeFlags = 1;
+use crate::src::looprestoration::LrEdgeFlags;
+
+
+
+
 #[derive(Copy, Clone)]
 #[repr(C)]
 pub union LooprestorationParams {
