+++ src/looprestoration.rs
@@ -1,0 +1,5 @@
+pub type LrEdgeFlags = libc::c_uint;
+pub const LR_HAVE_BOTTOM: LrEdgeFlags = 8;
+pub const LR_HAVE_TOP: LrEdgeFlags = 4;
+pub const LR_HAVE_RIGHT: LrEdgeFlags = 2;
+pub const LR_HAVE_LEFT: LrEdgeFlags = 1;
