+++ src/looprestoration_tmpl_8.rs
@@ -21,11 +21,11 @@ extern "C" {
 
 pub type pixel = uint8_t;
 pub type coef = int16_t;
-pub type LrEdgeFlags = libc::c_uint;
-pub const LR_HAVE_BOTTOM: LrEdgeFlags = 8;
-pub const LR_HAVE_TOP: LrEdgeFlags = 4;
-pub const LR_HAVE_RIGHT: LrEdgeFlags = 2;
-pub const LR_HAVE_LEFT: LrEdgeFlags = 1;
+use crate::src::looprestoration::LrEdgeFlags;
+use crate::src::looprestoration::LR_HAVE_BOTTOM;
+use crate::src::looprestoration::LR_HAVE_TOP;
+use crate::src::looprestoration::LR_HAVE_RIGHT;
+use crate::src::looprestoration::LR_HAVE_LEFT;
 pub type const_left_pixel_row = *const [pixel; 4];
 #[derive(Copy, Clone)]
 #[repr(C)]
