+++ src/recon_tmpl_16.rs
@@ -1248,11 +1248,11 @@ pub type looprestorationfilter_fn = Option::<
         libc::c_int,
     ) -> (),
 >;
-pub type LrEdgeFlags = libc::c_uint;
-pub const LR_HAVE_BOTTOM: LrEdgeFlags = 8;
-pub const LR_HAVE_TOP: LrEdgeFlags = 4;
-pub const LR_HAVE_RIGHT: LrEdgeFlags = 2;
-pub const LR_HAVE_LEFT: LrEdgeFlags = 1;
+use crate::src::looprestoration::LrEdgeFlags;
+
+
+
+
 #[derive(Copy, Clone)]
 #[repr(C)]
 pub union LooprestorationParams {
